Decouple neoServer creation in InProcessServer
- Create AbstractInProcessServerBuilder with all necessary functionality
- Add abstract method to create neoServer
- Create CommunityNeoServer in AbstractInProcessServerBuilder
  implementation
- Fix several formatting issues
