+++ community/neo4j-harness/src/main/java/org/neo4j/harness/internal/MapConfigurator.java
@@ -54,6 +54,4 @@ public class MapConfigurator implements ConfigurationBuilder
                 ConfiguratorWrappingConfigurationBuilder.toStringForThirdPartyPackageProperty( extensions ) );
         return new Config( serverConfigParams );
     }
-
-
 }
